requirements.txt
torch
torchvision
pandas
scikit-learn
numpy
pillow
tqdm
nltk
rouge-score
kagglehub
matplotlib
seaborn
scipy
plotly
pycocoevalcap
